requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
flower==2.0.1
humanize==4.12.3
kombu==5.5.4
packaging==25.0
pillow==11.2.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13